+++ src/lib/libc/stdio/swprintf.c
@@ -577,7 +577,7 @@ _DEFUN(_swprintf_r, (ptr, str, size, fmt),
   if (size > 0)  {
     *(wchar_t *)f._p = L'\0';	/* terminate the string */
   }
-  if(ret >= size)  {
+  if((size_t)ret >= size)  {
     /* _svfwprintf_r() returns how many wide characters it would have printed
      * if there were enough space.  Return an error if too big to fit in str,
      * unlike snprintf, which returns the size needed.  */
@@ -618,7 +618,7 @@ _DEFUN(swprintf, (str, size, fmt),
   if (size > 0)  {
     *(wchar_t *)f._p = L'\0';	/* terminate the string */
   }
-  if(ret >= size)  {
+  if((size_t)ret >= size)  {
     /* _svfwprintf_r() returns how many wide characters it would have printed
      * if there were enough space.  Return an error if too big to fit in str,
      * unlike snprintf, which returns the size needed.  */
